# [feature](bangc-ops): improve box_iou_rotated.