… docs/bangc-docs/design_docs/box_iou_rotated/res/cosine1.png
… docs/bangc-docs/design_docs/box_iou_rotated/res/sort_vertices1.png