Thanks for your contribution and we appreciate it a lot. :rocket::rocket:

## 1. Motivation

Improve the performance on 500.

## 2. Modification

modified:   bangc-ops/kernels/box_iou_rotated/box_iou_rotated_aligned.h
modified:   bangc-ops/kernels/box_iou_rotated/box_iou_rotated_nonaligned.h
modified:   bangc-ops/kernels/box_iou_rotated/box_iou_rotated_utils.h
modified:   bangc-ops/kernels/box_iou_rotated/box_iou_rotated.cpp
modified:   docs/bangc-docs/design_docs/box_iou_rotated/box_iou_rotated.md
new file:   docs/bangc-docs/design_docs/box_iou_rotated/res/cosine1.png
new file:   docs/bangc-docs/design_docs/box_iou_rotated/res/sort_vertices1.png

## 3. Test Report

If you want to know how to do operator testing, you can see [GTest-User-Guide-zh](https://github.com/Cambricon/mlu-ops/blob/master/docs/GTest-User-Guide-zh.md).

### 3.1 Modification Details

#### 3.1.1 Accuracy Acceptance Standard

For static threshold standard details, see: [MLU-OPS Accuracy Acceptance Standard](https://github.com/Cambricon/mlu-ops/blob/master/docs/MLU-OPS-Accuracy-Acceptance-Standard.md).

- static threshold
  - diff1
    - [ ] float32 mlu diff1 <= 1e-5
    - [x] float32 mlu diff1 <= 3e-3
    - [ ] float16 mlu diff1 <= 3e-3
  - diff2
    - [ ] float32 mlu diff2 <= 1e-5
    - [x] float32 mlu diff2 <= 3e-3
    - [ ] float16 mlu diff2 <= 3e-3
  - diff3
    - [ ] mlu diff3 == 0
    - [ ] mlu diff3_1 == 0
    - [ ] mlu diff3_2 == 0
- dynamic threshold
  - [ ] diff1: mlu diff1 <= max(baseline diff1 * 10, static threshold)
  - [ ] diff2: mlu diff2 <= max(baseline diff2 * 10, static threshold)
  - [ ] diff3: mlu diff3 <= max(baseline diff3 * 10, static threshold)
    - float32, threshold = 1e-5
    - float16, threshold = 1e-3

#### 3.1.2 Operator Scheme checklist

- Supported hardware
  - [x] MLU370
  - [x] MLU590
- Job types
  - [ ] BLOCK
  - [x] UNION1
  - [ ] UNION2
  - [ ] UNION4
  - [ ] The operator will dynamically select the most suitable task type, for example, UNION8

### 3.2 Accuracy Test

#### 3.2.1 Accuracy Test
If you have checked the following items, please tick the relevant box.

- [ ] Data type test (e.g. float32/int8)
- [ ] Multi-dimensional tensor test
- [ ] Layout test
- [ ] Different size/integer remainder end segment/alignment misalignment test
- [ ] Zero dimensional tensor test/zero element test
- [x] stability test
- [x] Multiple platform test
- [ ] Gen_case module test, see: [Gencase-User-Guide-zh](https://github.com/Cambricon/mlu-ops/blob/master/docs/Gencase-User-Guide-zh.md)
- [ ] Nan/INF tests 
- [ ] Bug fix tests
- [x] For memory leak check details, see: [GTest-User-Guide-zh](https://github.com/Cambricon/mlu-ops/blob/master/docs/GTest-User-Guide-zh.md)
- [x] For code coverage check details, see: [GTest-User-Guide-zh](https://github.com/Cambricon/mlu-ops/blob/master/docs/GTest-User-Guide-zh.md)
- [ ] For I/O calculation efficiency check details, see: [MLU-OPS-Performance-Acceptance-Standard](https://github.com/Cambricon/mlu-ops/blob/master/docs/MLU-OPS-Performance-Acceptance-Standard.md)

all release_test/release_temp/benchmark passed on 300 & 500.

#### 3.2.2 Parameter Check

Test Point-1: `MLU300 box num should less than 10,000,000`. Acceptance Standard: `Normal error`.
```bash
[2023-10-18 15:0:47] [MLUOP] [Warning]:Check failed: mt->shape_count == shape_count. "Parser: found shape count is not equal to value " "size.(mt->shape_count is value_size)"
[2023-10-18 15:0:48] [MLUOP] [Error]:[mluOpBoxIouRotated] Check failed: If it is aligned mode, on MLU300 box1_desc->dims[0] should less than or equal to 10,000,000 . But now is 429496730.
[2023-10-18 15:0:48] [MLUOP] [Error]:"MLUOP_STATUS_BAD_PARAM in mluOpBoxIouRotated(handle_, mode, aligned, box1, dev_box1, box2, dev_box2, ious, dev_ious)"
[2023-10-18 15:0:48] [MLUOP] [Warning]:Executor: reset cnrt and go on running, this may caused by cnrt failed.
/mlu-ops-dev/mlu-ops/bangc-ops/test/mlu_op_gtest/pb_gtest/src/gtest/mlu_op_gtest.cpp:125: Failure
```

Test Point-2: `MLU500 box tensor should less than 2G nums`. Acceptance Standard: `Normal error`.
```bash
[2023-10-18 14:34:4] [MLUOP] [Error]:[mluOpBoxIouRotated] overflow max supported tensor num 2147483647, now tensor's total num is 2147483650. 
[2023-10-18 14:34:4] [MLUOP] [Error]:"MLUOP_STATUS_NOT_SUPPORTED in mluOpBoxIouRotated(handle_, mode, aligned, box1, dev_box1, box2, dev_box2, ious, dev_ious)"
[2023-10-18 14:34:4] [MLUOP] [Warning]:Executor: reset cnrt and go on running, this may caused by cnrt failed.
/projs/platform/liuduanhui/mluops-dev/mlu-ops/bangc-ops/test/mlu_op_gtest/pb_gtest/src/gtest/mlu_op_gtest.cpp:125: Failure
```

### 3.3 Performance Test

See [MLU-OPS Performance Acceptance Standard](https://github.com/Cambricon/mlu-ops/blob/master/docs/MLU-OPS-Performance-Acceptance-Standard.md) for details.

Platform：MLU590

|         shape                | baseline(us) |  current(us)  | percent of <br>time reduced |
|------------------------------|----------|--------|--------|
|[32,5] aligned=1          | 82       | 24     | 70.73% |
|[640,5]  aligned=1         | 227.9    | 35     | 84.64% |
|[2400,5]  aligned=1        | 1082.05  | 34.44  | 96.82% |
|[18430,5] aligned=1       | 22225.76 | 440.75 | 98.02% |
|[38000,5] aligned=1       | 12217.35 | 246.38 | 97.98% |
|[160,5]  aligned=1         | 165      | 34     | 79.39% |
|[260,5]  aligned=1         | 108      | 29     | 73.15% |
|[570,5] aligned=1         | 190.31   | 31     | 83.71% |
|[2048,5] aligned=1        | 176.02   | 26     | 85.23% |
|[18432,5] aligned=1       | 599.04   | 68     | 88.65% |
|[250,5],[5,5] aligned=0   | 2014.03  | 1877.01| 6.80%  |
|[60,5],[30,5]  aligned=0   | 1909.06  | 697.82 | 63.45% |
|[10,5],[30,5]  aligned=0   | 328      | 208    | 36.59% |
| [5,5],[320,5] aligned=0  | 828      | 46     | 94.44% |

### 3.4 Summary Analysis

The performance improve by 6.8 % to 98% on 500.